backend/blockchain/requirements.txt
# Web3 for blockchain interaction
web3>=6.0.0

# Environment variables
python-dotenv>=1.0.0

# Ethereum account management
eth-account>=0.11.0



